backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.48
aiosqlite==0.20.0
asyncpg==0.30.0
redis==5.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
google-generativeai==0.7.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
structlog==24.1.0
Pillow==10.4.0
pytesseract==0.3.10
PyMuPDF==1.24.10
boto3==1.37.22
